Update typings to reflect moving `fileTypeFromBlob()` from browser to core.
