+++ core.d.ts
@@ -401,3 +401,21 @@ if (stream2.fileType?.mime === 'image/jpeg') {
 ```
 */
 export function fileTypeStream(readableStream: ReadableStream, options?: StreamOptions): Promise<ReadableStreamWithFileType>;
+
+/**
+Detect the file type of a [`Blob`](https://nodejs.org/api/buffer.html#class-blob).
+
+@example
+```
+import {fileTypeFromBlob} from 'file-type';
+
+const blob = new Blob(['<?xml version="1.0" encoding="ISO-8859-1" ?>'], {
+	type: 'plain/text',
+	endings: 'native'
+});
+
+console.log(await fileTypeFromBlob(blob));
+//=> {ext: 'txt', mime: 'plain/text'}
+```
+ */
+export declare function fileTypeFromBlob(blob: Blob): Promise<FileTypeResult | undefined>;
